requirements.txt
python-telegram-bot==22.6
notion-client==3.0.0
python-dotenv==1.2.2